requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==10.4.0
pillow_heif==0.18.0
Pygments==2.18.0
rich==13.7.1
